fig: float to int

+++ pyrpl/software_modules/loop.py
@@ -56,7 +56,7 @@ class Loop(Module):
 
     @interval.setter
     def interval(self, val):
-        self.timer.setInterval(val*1000.0)
+        self.timer.setInterval(int(val*1000.0))
 
     def _clear(self):
         self._ended = True
